processed comments code review replaced: NotImplementedError with warning added: tests refactored: ArgillaTrainingMixin refactored: ArgillaUnificationMixin
